refactored the Decapolis solution to use an Onion Architecture at least as far as showing the structure for separation of concerns.  Core/Presentation/Infrastructure solution folders were added.  Inside Core, there is a Domain project and an Application project.  The Application project shows a query using a CQRS pattern.  AutoMapper was utilized to have request/response mappings for the API models to/from Domain Entities and requests handled by MediatR.
If additional architectural improvements are needed to show separation of concerns, I could extend the solution to add an IApplicationDbContext in the Application Layer and do a EFCore implementation of that in the Infrastucture Layer.  Currently, I am just filling up a list of cities in code in the Application layer which obviously is not would be done in a real world application.